Rename kind/depracation folder to kind/deprecations
