+++ lib/kind/deprecations/built_in_type_checkers.rb
