+++ lib/kind/deprecations/checker/factory.rb
