+++ lib/kind/deprecations/checker/protocol.rb
