+++ lib/kind/deprecations/is.rb
