+++ lib/kind/deprecations/of.rb
