+++ lib/kind/deprecations/types.rb
